# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/questions:
    get:
      tags:
        - Questions
      summary: List Questions
      description: Get a paginated list of Questions with optional filtering and sorting
      operationId: listQuestions
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Questions
      summary: Create Question
      description: Create a new Question
      operationId: createQuestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuestionRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/question_answers:
    get:
      tags:
        - QuestionAnswers
      summary: List QuestionAnswers
      description: Get a paginated list of QuestionAnswers with optional filtering and sorting
      operationId: listQuestionAnswers
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionAnswerList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - QuestionAnswers
      summary: Create QuestionAnswer
      description: Create a new QuestionAnswer
      operationId: createQuestionAnswer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuestionAnswerRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionAnswer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    Question:
      type: object
      required:
        - id
        - survey_id
        - sequence
        - is_page
        - title
        - random_questions_count
        - is_scored_question
        - answer_score
        - save_as_email
        - save_as_nickname
        - matrix_subtype
        - scale_min
        - scale_max
        - is_time_limited
        - is_time_customized
        - comments_allowed
        - comment_count_as_answer
        - validation_required
        - validation_email
        - validation_length_min
        - validation_length_max
        - constr_mandatory
        - metadata
      properties:
        id:
          type: string
          format: uuid
        survey_id:
          type: string
          format: uuid
        sequence:
          type: integer
          format: int32
        is_page:
          type: boolean
        question_type:
          nullable: true
          type: string
        title:
          type: string
          maxLength: 255
        description:
          nullable: true
          type: string
        question_placeholder:
          nullable: true
          type: string
        background_image:
          nullable: true
          type: string
        random_questions_count:
          type: integer
          format: int32
        is_scored_question:
          type: boolean
        answer_numerical_box:
          nullable: true
          type: number
          format: float
        answer_date:
          nullable: true
          type: string
          format: date
        answer_datetime:
          nullable: true
          type: string
          format: date-time
        answer_score:
          type: number
          format: float
        save_as_email:
          type: boolean
        save_as_nickname:
          type: boolean
        matrix_subtype:
          type: string
        scale_min:
          type: integer
          format: int32
        scale_max:
          type: integer
          format: int32
        scale_min_label:
          nullable: true
          type: string
        scale_mid_label:
          nullable: true
          type: string
        scale_max_label:
          nullable: true
          type: string
        is_time_limited:
          type: boolean
        time_limit:
          nullable: true
          type: integer
          format: int32
        is_time_customized:
          type: boolean
        comments_allowed:
          type: boolean
        comments_message:
          nullable: true
          type: string
        comment_count_as_answer:
          type: boolean
        validation_required:
          type: boolean
        validation_email:
          type: boolean
        validation_length_min:
          type: integer
          format: int32
        validation_length_max:
          type: integer
          format: int32
        validation_min_float_value:
          nullable: true
          type: number
          format: float
        validation_max_float_value:
          nullable: true
          type: number
          format: float
        validation_min_date:
          nullable: true
          type: string
          format: date
        validation_max_date:
          nullable: true
          type: string
          format: date
        validation_min_datetime:
          nullable: true
          type: string
          format: date-time
        validation_max_datetime:
          nullable: true
          type: string
          format: date-time
        validation_error_msg:
          nullable: true
          type: string
        constr_error_msg:
          nullable: true
          type: string
        constr_mandatory:
          type: boolean
        page_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateQuestionRequest:
      type: object
      required:
        - survey_id
        - sequence
        - is_page
        - title
        - random_questions_count
        - is_scored_question
        - answer_score
        - save_as_email
        - save_as_nickname
        - matrix_subtype
        - scale_min
        - scale_max
        - is_time_limited
        - is_time_customized
        - comments_allowed
        - comment_count_as_answer
        - validation_required
        - validation_email
        - validation_length_min
        - validation_length_max
        - constr_mandatory
        - metadata
      properties:
        survey_id:
          type: string
          format: uuid
        sequence:
          type: integer
          format: int32
        is_page:
          type: boolean
        question_type:
          nullable: true
          type: string
        title:
          type: string
          maxLength: 255
        description:
          nullable: true
          type: string
        question_placeholder:
          nullable: true
          type: string
        background_image:
          nullable: true
          type: string
        random_questions_count:
          type: integer
          format: int32
        is_scored_question:
          type: boolean
        answer_numerical_box:
          nullable: true
          type: number
          format: float
        answer_date:
          nullable: true
          type: string
          format: date
        answer_datetime:
          nullable: true
          type: string
          format: date-time
        answer_score:
          type: number
          format: float
        save_as_email:
          type: boolean
        save_as_nickname:
          type: boolean
        matrix_subtype:
          type: string
        scale_min:
          type: integer
          format: int32
        scale_max:
          type: integer
          format: int32
        scale_min_label:
          nullable: true
          type: string
        scale_mid_label:
          nullable: true
          type: string
        scale_max_label:
          nullable: true
          type: string
        is_time_limited:
          type: boolean
        time_limit:
          nullable: true
          type: integer
          format: int32
        is_time_customized:
          type: boolean
        comments_allowed:
          type: boolean
        comments_message:
          nullable: true
          type: string
        comment_count_as_answer:
          type: boolean
        validation_required:
          type: boolean
        validation_email:
          type: boolean
        validation_length_min:
          type: integer
          format: int32
        validation_length_max:
          type: integer
          format: int32
        validation_min_float_value:
          nullable: true
          type: number
          format: float
        validation_max_float_value:
          nullable: true
          type: number
          format: float
        validation_min_date:
          nullable: true
          type: string
          format: date
        validation_max_date:
          nullable: true
          type: string
          format: date
        validation_min_datetime:
          nullable: true
          type: string
          format: date-time
        validation_max_datetime:
          nullable: true
          type: string
          format: date-time
        validation_error_msg:
          nullable: true
          type: string
        constr_error_msg:
          nullable: true
          type: string
        constr_mandatory:
          type: boolean
        page_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    QuestionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Question'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    QuestionAnswer:
      type: object
      required:
        - id
        - sequence
        - is_correct
        - answer_score
        - metadata
      properties:
        id:
          type: string
          format: uuid
        question_id:
          nullable: true
          type: string
          format: uuid
        matrix_question_id:
          nullable: true
          type: string
          format: uuid
        sequence:
          type: integer
          format: int32
        value:
          nullable: true
          type: string
        value_image:
          nullable: true
          type: string
        value_image_filename:
          nullable: true
          type: string
        is_correct:
          type: boolean
        answer_score:
          type: number
          format: float
        metadata:
          type: object
...
    CreateQuestionAnswerRequest:
      type: object
      required:
        - sequence
        - is_correct
        - answer_score
        - metadata
      properties:
        question_id:
          nullable: true
          type: string
          format: uuid
        matrix_question_id:
          nullable: true
          type: string
          format: uuid
        sequence:
          type: integer
          format: int32
        value:
          nullable: true
          type: string
        value_image:
          nullable: true
          type: string
        value_image_filename:
          nullable: true
          type: string
        is_correct:
          type: boolean
        answer_score:
          type: number
          format: float
        metadata:
          type: object
...
    QuestionAnswerList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/QuestionAnswer'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []